headless: Rename StubHost to HeadlessHost.
It's had that weird name forever, get rid of it.
