+++ headless/Headless.vcxproj
@@ -477,7 +477,7 @@
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|ARM64'">true</ExcludedFromBuild>
       <ExcludedFromBuild Condition="'$(Configuration)|$(Platform)'=='Release|ARM'">true</ExcludedFromBuild>
     </ClCompile>
-    <ClCompile Include="StubHost.cpp" />
+    <ClCompile Include="HeadlessHost.cpp" />
     <ClCompile Include="WindowsHeadlessHost.cpp" />
   </ItemGroup>
   <ItemGroup>
@@ -510,7 +510,7 @@
   <ItemGroup>
     <ClInclude Include="Compare.h" />
     <ClInclude Include="SDLHeadlessHost.h" />
-    <ClInclude Include="StubHost.h" />
+    <ClInclude Include="HeadlessHost.h" />
     <ClInclude Include="WindowsHeadlessHost.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
